SQLite: create a function that reports if a rune is alphabetic
Do this because this way we can reuse the code and avoid repeating
it in several places.

Use the definition of alphabetical given in the SQLite documentation.
Note that any rune greater than 0x7F is considered alphabetic by the
documentation, so do not use unicode.IsLetter because there are runes
greater than 0x7F that are not letters.