SQLite: use the new method peekRune
The Next method reads the next rune and if it is for the beginning of
a word it calls unreadRune so that the word method reads the first
rune of the word.

Use the new peekRune method so that there is no need to call unreadRune,
because this simplifies the code.