SQLite: fix: correct the word method to handle appropriately pairs of double-quotes and pairs of grave accents
For the double quoted identifier the word method stops reading at the
first double-quote, ignoring the possibility of a pair of double quotes.
Correct this problem.

A similar problem occurs for the case of grave accented identifiers,
but with respect to the possibility of a pair of grave accents.
Correct this problem.