SQLite: make the lexer scan white spaces instead of discarding them
Commit 118a1a7 (SQLite: add the method that discards white space on the lexer, 2024-09-11) noticed that
the lexer could discard whitespaces. But we came to the conclusion that one of the formatters will need the spaces.
We are going to make a formatter that uses only lexical information, so it will not be possible to indent or move
the white spaces, so this formatter will need the spaces to know where to place the lexemes.

Because of this, create a type of token for white spaces and make the lexer scan them.